requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-openai>=0.1.0
langchain-google-genai>=0.1.0
langchain-community>=0.1.0
langchain-text-splitters>=0.1.0
langchain-core>=0.1.0
chromadb>=0.4.22
pypdf>=3.17.0
python-dotenv>=1.0.0
asyncio>=3.4.3 